03-build/mac-agent/requirements.txt
pyobjc-framework-Contacts>=10.0
requests>=2.32
